cleanup part 1

+++ OTVision/track/track.py
@@ -1,7 +1,9 @@
-# OTVision: Python module to calculate homography matrix from reference
+"""
+OTVision main module for tracking objects in successive frames of videos.
+"""
 # points and transform tracksectory points from pixel into world coordinates.
 
-# Copyright (C) 2020 OpenTrafficCam Contributors
+# Copyright (C) 2022 OpenTrafficCam Contributors
 # <https://github.com/OpenTrafficCam
 # <team@opentrafficcam.org>
 #
@@ -20,14 +22,11 @@
 
 
 import json
-import logging
 from pathlib import Path
-
-import geopandas as gpd
-import pandas as pd
 
 from OTVision.config import CONFIG
 from OTVision.helpers.files import denormalize, get_files
+from OTVision.helpers.log import log
 
 from .iou import track_iou
 
@@ -40,20 +39,26 @@ def main(
     sigma_iou=CONFIG["TRACK"]["IOU"]["SIGMA_IOU"],
     t_min=CONFIG["TRACK"]["IOU"]["T_MIN"],
     t_miss_max=CONFIG["TRACK"]["IOU"]["T_MISS_MAX"],
-    overwrite=CONFIG["TRACK"]["IOU"]["OVERWRITE"],
+    overwrite=CONFIG["TRACK"]["OVERWRITE"],
+    debug: bool = CONFIG["TRACK"]["DEBUG"],
 ):
-    filetypes = CONFIG["DEFAULT_FILETYPE"]["DETECT"]
-    detections_files = get_files(paths, filetypes)
+    log.info("Start tracking")
+    if debug:
+        log.setLevel("DEBUG")
+        log.debug("Logging track.main() in debug mode")
+
+    filetype = CONFIG["DEFAULT_FILETYPE"]["DETECT"]
+    detections_files = get_files(paths, filetype)
     for detections_file in detections_files:
-        logging.info(f"New detections file: {detections_file}")
+        log.info(f"Try tracking {detections_file}")
 
         try:
             with open(detections_file) as f:
                 detections = json.load(f)
-            logging.info("detections read")
+            log.info(f"{filetype} read")
 
             detections_denormalized = denormalize(detections)
-            logging.info("detections denormalized")
+            log.info("Detections denormalized")
 
             tracks_px, trajectories_geojson = track(
                 detections=detections_denormalized,
@@ -64,23 +69,23 @@ def main(
                 t_min=t_min,
                 t_miss_max=t_miss_max,
             )
-            logging.info("detections tracked")
+            log.info("Detections tracked")
 
             write(
                 tracks_px=tracks_px,
                 trajectories_geojson=trajectories_geojson,
                 detections_file=detections_file,
+                overwrite=overwrite,
             )
-            logging.info("Tracking finished")
         except OSError as oe:
-            logging.error(
+            log.error(
                 (
                     f'Could not open "{detections_file}". '
                     f"Following exception occured: {str(oe)}"
                 )
             )
         except json.JSONDecodeError as je:
-            logging.error(
+            log.error(
                 (
                     f'Unable to decode "{detections_file}" as JSON.'
                     f"Following exception occured: {str(je)}"
@@ -105,19 +110,23 @@ def track(
         t_min=t_min,
         t_miss_max=t_miss_max,
     )
-    trk_config = {}
-    trk_config["yolo_mode"] = yolo_mode
-    trk_config["tracker"] = "IOU"
-    trk_config["sigma_l"] = sigma_l
-    trk_config["sigma_h"] = sigma_h
-    trk_config["sigma_iou"] = sigma_iou
-    trk_config["t_min"] = t_min
-    trk_config["t_miss_max"] = t_miss_max
-    tracks_px = {}
-    tracks_px["vid_config"] = detections["vid_config"]
-    tracks_px["det_config"] = detections["det_config"]
-    tracks_px["trk_config"] = trk_config
-    tracks_px["data"] = new_detections
+
+    trk_config = {
+        "yolo_mode": yolo_mode,
+        "tracker": "IOU",
+        "sigma_l": sigma_l,
+        "sigma_h": sigma_h,
+        "sigma_iou": sigma_iou,
+        "t_min": t_min,
+        "t_miss_max": t_miss_max,
+    }
+
+    tracks_px = {
+        "vid_config": detections["vid_config"],
+        "det_config": detections["det_config"],
+        "trk_config": trk_config,
+        "data": new_detections,
+    }
 
     return tracks_px, trajectories_geojson
 
@@ -125,78 +134,20 @@ def track(
 # TODO: Implement overwrite as in detect, maybe refactor?
 def write(
     tracks_px,
-    trajectories_geojson,
     detections_file,
     overwrite=CONFIG["TRACK"]["IOU"]["OVERWRITE"],
 ):
-    # TODO: #96 Make writing each filetype optional (list of filetypes as parameter)
-    # Write JSON
-    with open(Path(detections_file).with_suffix(".ottrk"), "w") as f:
-        json.dump(tracks_px, f, indent=4)
-    logging.info("JSON written")
-
-    # Write GeoJSON
-    # TODO: #95 Add metadata to GeoJSON (and GPKG)
-    with open(Path(detections_file).with_suffix(".GeoJSON"), "w") as f:
-        json.dump(trajectories_geojson, f, indent=4)
-    logging.info("GeoJSON written")
-
-    # Convert to geodataframe and write GPKG
-    get_geodataframe_from_framewise_tracks(tracks_px).to_file(
-        Path(detections_file).with_suffix(".gpkg"), driver="GPKG"
-    )
-    logging.info("GPKG written")
-
-
-def get_geodataframe_from_framewise_tracks(tracks_px):
-    # TODO: #93 Create LineString GeoDataFrame from Point GeoDataFrame
-    df_trajectories = pd.DataFrame.from_dict(
-        {
-            (frame_nr, det_nr): tracks_px["data"][frame_nr][det_nr]
-            for frame_nr in tracks_px["data"].keys()
-            for det_nr in tracks_px["data"][frame_nr].keys()
-        },
-        orient="index",
-    ).rename_axis(("frame", "ID"))
-    gdf_trajectories = gpd.GeoDataFrame(
-        df_trajectories,
-        geometry=gpd.points_from_xy(df_trajectories.x, df_trajectories.y),
-    )
-
-    # aggregate these points with the GrouBy
-    # gdf_trajectories = (
-    #     (gdf_trajectories.swaplevel().sort_index())
-    #     .groupby(["ID"])
-    #     .agg(
-    #         {
-    #             "class": pd.Series.mode,
-    #             "conf": "max",
-    #             "geometry": pd.Series.apply(
-    #                 self=gdf_trajectories["geometry"],
-    #                 # BUG: #97 Cannot convert Point to list
-    #                 func=lambda x: LineString(x.tolist()),
-    #             ),
-    #         }
-    #     )  # ["geometry"]
-    #    .size()  # apply(lambda x: LineString(x.tolist()))
-    # )
-    # gdf_trajectories = gpd.GeoDataFrame(gdf_trajectories, geometry="geometry")
-    print(gdf_trajectories)
-
-    return gdf_trajectories
-
-
-def get_geodataframe_from_trackwise_tracks(trajectories_px):
-    # TODO: #94 Directly create LineString GeoDataFrame
-    df_trajectories = pd.DataFrame(trajectories_px)
-    df_trajectories = df_trajectories.df.drop(["frames", "bboxes", "conf"], axis=1)
-    # gdf_trajectories = gpd.GeoDataFrame(
-    #     df_trajectories,
-    #     geometry=gpd.points_from_xy(df_trajectories.x, df_trajectories.y),
-    # )
-    return df_trajectories  # gdf_trajectories
-
-
-# To Dos:
-# - Improve memory management (MemoryError after two 2h detections)
-# - Perform logging by dedicated python package
+    # ?: Check overwrite before tracking instead of before writing tracking?
+    # TODO: Export also as csv, trj and alternative json
+    tracks_file = Path(detections_file).with_suffix(CONFIG["FILETYPES"]["TRACK"])
+    tracks_file_already_exists = tracks_file.is_file()
+    if overwrite or not tracks_file_already_exists:
+        # Write JSON
+        with open(tracks_file, "w") as f:
+            json.dump(tracks_px, f, indent=4)
+        if tracks_file_already_exists:
+            log.info(f"{tracks_file} overwritten")
+        else:
+            log.info(f"{tracks_file}  file written")
+    else:
+        log.info(f"{tracks_file} already exists. To overwrite, set overwrite=True")
